mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.2",
    "fastmcp>=1.0.0",
    "google-genai>=1.50.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
